+++ edenai_apis/apis/microsoft/microsoft_ocr_api.py
@@ -202,7 +202,7 @@ class MicrosoftOcrApi(OcrInterface):
                 description = item["value"].get("Name", default_dict).get("value")
                 price = item["value"].get("Price", default_dict).get("value")
                 quantity_str = item["value"].get("Quantity", default_dict).get("value")
-                quantity = int(quantity_str) if quantity_str else None
+                quantity = float(quantity_str) if quantity_str else None
                 total = item["value"].get("TotalPrice", default_dict).get("value")
                 items.append(
                     ItemLines(
